+++ src/components/index.js
@@ -1,0 +1,9 @@
+export {Hamerburger} from './svgs/hamerburger';
+export {Camera} from './svgs/camera';
+export {BorderHeart} from './svgs/borderheart';
+export {CreateRecipe} from './svgs/createrecipe';
+export {CreateRecipeWhite} from './svgs/createrecipewhite';
+export {MainCamera} from './svgs/maincamera';
+export {Search} from './svgs/search';
+export {ThreeDots} from './svgs/threedots';
+export {Tips } from './svgs/tips';
