+++ src/components/svgs/camera.js
@@ -1,0 +1,34 @@
+export const Camera = () => {
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="20"
+        height="20"
+        fill="none"
+        viewBox="0 0 20 20"
+        class="loading:hidden mise-icon mise-icon-camera"
+      >
+        <path
+          stroke="currentColor"
+          stroke-width="1.3"
+          d="M15 17.5H5a3.333 3.333 0 0 1-3.333-3.333v-7.5c0-.92.746-1.667 1.667-1.667h1.608c.557 0 1.077-.279 1.386-.742l.677-1.016c.31-.463.83-.742 1.387-.742h3.216c.557 0 1.078.279 1.387.742l.677 1.016c.31.463.83.742 1.387.742h1.608c.92 0 1.666.746 1.666 1.667v7.5A3.333 3.333 0 0 1 15 17.5Z"
+        ></path>
+        <circle
+          cx="4.167"
+          cy="7.5"
+          r=".833"
+          fill="currentColor"
+          stroke="currentColor"
+          stroke-width=".1"
+          transform="rotate(-180 4.167 7.5)"
+        ></circle>
+        <circle
+          cx="10"
+          cy="10.834"
+          r="2.5"
+          stroke="currentColor"
+          stroke-width="1.3"
+        ></circle>
+      </svg>
+    );
+  };
