+++ src/components/svgs/search.js
@@ -1,0 +1,25 @@
+export const Search = () => {
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="24"
+        height="24"
+        fill="none"
+        viewBox="0 0 24 24"
+        class="text-cookpad-gray-500 mr-sm mise-icon mise-icon-search-unselected"
+      >
+        <path
+          stroke="currentColor"
+          stroke-width="1.5"
+          d="M17 10a7 7 0 1 1-14 0 7 7 0 0 1 14 0Z"
+        ></path>
+        <path
+          stroke="currentColor"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+          stroke-width="1.5"
+          d="m15 15 5 5"
+        ></path>
+      </svg>
+    );
+  };
