+++ src/components/svgs/threedots.js
@@ -1,0 +1,21 @@
+export const ThreeDots = () => {
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="24"
+        height="24"
+        fill="none"
+        viewBox="0 0 24 24"
+        class="text-cookpad-gray-400 mise-icon mise-icon-overflow-horizontal"
+      >
+        <path
+          fill="currentColor"
+          stroke="currentColor"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+          stroke-width="1.5"
+          d="M11 12a1 1 0 1 0 2 0 1 1 0 0 0-2 0ZM18 12a1 1 0 1 0 2 0 1 1 0 0 0-2 0ZM4 12a1 1 0 1 0 2 0 1 1 0 0 0-2 0Z"
+        ></path>
+      </svg>
+    );
+  };
