+++ src/components/svgs/tips.js
@@ -1,0 +1,27 @@
+export const Tips = () => {
+    return (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="20"
+        height="20"
+        fill="none"
+        viewBox="0 0 20 20"
+        class="btn__icon mise-icon mise-icon-add-tip"
+        style={{ color: "white" }}
+      >
+        <path
+          stroke="currentColor"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+          stroke-width="1.3"
+          d="M16.666 1.667v5M14.166 4.167h5M7.5 14.166S8.75 15 10 15s2.5-.834 2.5-.834"
+        ></path>
+        <path
+          stroke="currentColor"
+          stroke-linecap="round"
+          stroke-width="1.3"
+          d="M12.169 2.083a5.833 5.833 0 0 0-6.402 9.43c.905.955 1.733 2.097 1.733 3.413v.907a2.5 2.5 0 0 0 2.5 2.5v0a2.5 2.5 0 0 0 2.5-2.5v-.907c0-1.316.827-2.458 1.733-3.412A5.826 5.826 0 0 0 15.6 9.133"
+        ></path>
+      </svg>
+    );
+  };
